Partial fix for liboai not finding dependent libraries
Prior commit was not finding the json lib nor the curl lib.
This is obviously because it is not built yet, however it is configured.
This points into the cmake build directory where the .cmake files are
generated, and adds those to the cmake search path.

This is still incomplete, however, because when finding curl, its find
CURLConfig.cmake is looking for CURLTargets.cmake - which I believe is
not generated until build time. Some hacks to try to generate it at
configure time have failed for me.